[550389] Consider shadow border size of containers in ELK layout
For container, except workspace images and regions, a shadow of 2 pixels
are added inside the figure bounds (thanks to
org.eclipse.sirius.diagram.ui.edit.api.part.AbstractDiagramElementContainerEditPart.addDropShadow(NodeFigure,
IFigure)). If an edge starts from a container with shadow, it starts
from the figure without the shadow. This constraint must be passed to
ELK to avoid to have wrong bendpoints after layout. It is the same for
east and south border nodes.
This commit removes the shadow border size during the conversion from
Sirius diagram to ELK graph and adds it again when applying ELK layout
to Sirius diagram.

Bug: 550389
Cherry-picked-from: 548820
Change-Id: I5d4a466862f26d65f5c4ee4239f3458f2297a6ce
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>