[549148] Remove BorderItemAwareLayoutProvider for CustomLayoutAlgorithm
If a specific layout, ie CustomLayoutAlgorithm, is specified in the VSM,
we consider that there is no additional work to do on border nodes.
Before this commit, with ELK for example, the border nodes were moved
with this BorderItemAwareLayoutProvider with a worst result.
The removed code was added for ELK in context of the "bugzilla 540987 -
[ELK] Arranging from selection and pinned elements not working" [1].

[1] https://bugs.eclipse.org/bugs/show_bug.cgi?id=540987

Bug: 549148
Change-Id: I9036aadc38128e7f3bec8886d66338d22cf40c78
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>