[550371] Move LocalResourceCollector under SessionLazyCrossReferencer
The LocalResourceCollector is now extending SessionLazyCrossReferencer.
The DASI now instantiates only one instance of
LocalResourceCollectorCrossReferencer
instead of  one instance of LocalResourceCollectorCrossReferencer and
one instance of SessionLazyCrossReferencer.

Some code has been moved back from SessionLazyCrossReferencer to
SiriusCrossReferenceAdapter as it does not depend on the lazy contract
but to one of the previous midclass which was merged into
SiriusCrossReferenceAdapter.

Bug: 550371
Change-Id: I08785a2efd77c0059fd46fe3070695a2072114a9
Signed-off-by: Pierre Guilet <pierre.guilet@obeo.fr>