[548631] Fix tests and bugs after name and description move
- Update DRepInDViewToRootObjectsAndWithDRepDescRepPathMigrationParticipant
to move description and name from DRepresentation to
DRepresentationDescriptor.
- Update metamodel to have a transient name and documentation EAttribute
on DRepresentation so aql:diagram.name and aql:diagram.documentation can
work. The value is computed from the value in DRepresentationDescriptor.
- Fix DocumentationPropertySection not working because it was trying to
set DRepresentation documentation attribute instead of the
DRepresentationDescriptor one.
-Fix change id not updated with notifications of multiple
representations and wrong change id update with notification coming from
an element which have one of its parent target of a transient feature.

Bug: 548631 
Change-Id: Ifcf4ee43af5e4440d0e6b79a79020ecfdb44ce5c
Signed-off-by: Pierre Guilet <pierre.guilet@obeo.fr>