[549887] Use a more precise label size estimation in GMFHelper
When computing a label size in GMFHelper, we always use {50, 20} but
sometimes, it is possible to evaluate more precisely the label size. In
case of "Invalid thread access", the previous dimension will be used.
This change is indirectly used by CanonicalDBorderItemLocator and allows
to have right location for label on border. It was false before but
"visually fixed" by Draw2D.
The code to compute a "more precisely label size" has been extracted
from org.eclipse.sirius.diagram.ui.internal.operation.RegionContainerUpdateLayoutOperation.

Bug: 549887
Change-Id: I075aa6bcee8ad47e71b895784dccbc3cc36a7bad
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>