[550371] Fix detection of transitive semantic resources
The resource collector is no more added on the whole resource set, its
result now depends on the semantic cross referencer installation state.

- The collectAllReferencedResources(res) is still transitive but only on
resource on which the semantic cross referencer is installed, the
doAddSemanticResource is responsible for the installation of the
semantic cross referencer is now recursive as the
collectAllReferencedResource, it is now recursive.

Bug: 550371
Change-Id: I957734a08662666675b14732bfe0bc384643f4a4
Signed-off-by: Maxime Porhel <maxime.porhel@obeo.fr>