[cleanup] Avoid unneeded modeling explorer viewer refresh
During test analysis, I found some unexpected refreshes of the view.
When a session is closed, the RefreshViewerTrigger is removed from the
resource set listeners and a refresh is done, see
org.eclipse.sirius.ui.tools.internal.views.common.navigator.SiriusCommonContentProvider.CommonSessionManagerListener.notifyRemoveSession(Session)

But between the session.close() and the removal of RefreshViewerTrigger,
we might have additional event which will try to update the view : an
analysis is done to detect if we can retrieve some elements to
update/refresh or if we need to refresh the viewer.

If the session is closed, this means that the trigger will be removed
soon and the refresh will be done on the view, then we can safely do
nothing is the sessionRefreshViewerTrigger.

Change-Id: I3f672ce436f28d6cb4d736a95dcfb0700364db2e
Signed-off-by: Maxime Porhel <maxime.porhel@obeo.fr>