[551053] Catch exceptions on session close with dirty state
When closing a session with a dirty state, a popup is opened to ask the
user if he want to save before closing. If the choice is to save, it is
saved as expected but some exception are thrown because some refresh are
triggered even though the session is already closed. These exception
should be silently catched.

Bug: 551053
Change-Id: Iebfd46cf5712dcd73a5071ec8519895e4a22cfe4
Signed-off-by: Steve Monnier <steve.monnier@obeo.fr>