[548427] Fix regression caused by the new ChangeIdUpdaterListener
In case of rollback, for example, some DRepresentationDescriptor can be
added but with a Change Id already set. In this situation, we must not
compute again the Change Id. This regression has been detected by a
automatic test from the "team part" of Sirius:
fr.obeo.dsl.viewpoint.tests.collab.swtbot.general.UndoAfterCommitTest.testUndoAfterCommitFailed.

This commit also documents an API change not documented in the previous
commit.

Bug: 548427
Change-Id: Ia6960c41b29132950ac31dff61bb4bce77400093
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>