[552747] Fix ELK "layout problem" after diagram instantiation
1- Consider AbstractDNode*CompartmentEditPart in
EditPartQuery.getDDiagram() : AbstractDNodeContainerCompartmentEditPart
and AbstractDNodeListCompartmentEditPart are not correctly managed by
EditPartQuery.getDDiagram() as they do not implement
IDiagramElementEditPart.

2- This commit also customizes the arrange called after the creation to
avoid NPE and/or layout problem with ELK algorithm. With the previous
code, several "arrange all" are called (one for the diagram, and one per
container). It seems OK with "classical layout". But it is problematic
for ELK layouts. So for ELK we keep only the first group of diagram
elements (child of diagram).

3- This commit also changes bounds of a container and its node from data
needed for EditModeTest. This test uses ELK, I don't know why, but after
this commit, the layout a the container is not done and in the rest of
test, we try to double-click on a node without success. Previously, a
"standard" layout, ie not an ELK one, is done on the contents of the
container.

This commit is a first in a series of patches to completely fix the
ticket. Indeed, it solves the identified problem but it seems not normal
that:
* several arranges are launched at opening.
* container contents are not arranged like it is the case for other
layout.

Bug: 552747
Change-Id: I6d485d2ae1e0a845b11b42a4b60ff7ca5488c5bd
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>