[test] Add test has unreliable because of HTML problem rendering
2 tests fail in MigrationTest class. These tests fail on some IC server
because of a problem of rendering HTML in the aird editor. A popup
appears with title "Internal Error" and message "An SWT error has
occurred. You are recommended to exit the workbench...". So the test
fails.
In org.eclipse.sirius.ui.tools.internal.viewpoint.ViewpointsSelectionGraphicalHandler.createBrowser(
 Composite), it seems that this case should be correctly handled but it
is not the case.

Change-Id: I88afa0f8968383fd5c29343069853bef2a48b497
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>