[552839] Catch new events to close the opened editor on deletion
It appears that the type of event that remove the target of
DRepresentationDescriptor might defers depending on several factors.
Depending, on the available bundles in the platform the removal of the
target from the DRepresentation descriptor is either done by unsetting
the feature (actual case) or by setting the feature to null. As we do
not control in which order the notifications and the operations are
done, we need to handle both events. This commit do not provide unit
test because the problem occurs only when some plugins (cdo) are part of
the running platform.


Bug: 552839
Change-Id: Ie8f050176394e16d4b930fce6fdf25527594c013
Signed-off-by: Arthur Daussy <arthur.daussy@obeo.fr>
Signed-off-by: Maxime Porhel <maxime.porhel@obeo.fr>