[553182] Prevents NPE notification regarding deleted representations
In some specific uses cases (Capella modeler for example), the
ResourceSet can send notifications on DRepresentations that have been
deleted (Post or pre commit listeners for example). In such case, the
DRepresentation is already removed from its resource making impossible
to find the session. Such cases should be guarded against NPE.

This problem can't be reproduced in pure Sirius environment. When adding
to the running platform (CDO, Collab or Collab) it can be reporuced in
some cases. For example, in Capella by deleting an ancestor of the
semantic target of an opened representation might raises some NPE.

In order to reproduce see bug:
https://bugs.polarsys.org/show_bug.cgi?id=2668
https://bugs.polarsys.org/show_bug.cgi?id=2669

Bug: 553182
Change-Id: Ia87239bf2c51e892b9271b4933c08d3434a85b59
Signed-off-by: Arthur Daussy <arthur.daussy@obeo.fr>
Signed-off-by: Maxime Porhel <maxime.porhel@obeo.fr>