[551235] Make Session1MillionTests timeouts configurable
Add support for system properties to override the maximum open & close
times without changing the code. Useful to tweak the values depending
on the system where the tests are run, for example on a slow/heavily
loaded CI server.

To use these settings:

  -Dorg.eclipse.sirius.tests.unit.perf.common.Session1MillionTests.maxOpenTimeInSeconds=30
  -Dorg.eclipse.sirius.tests.unit.perf.common.Session1MillionTests.maxCloseTimeInSeconds=20

Currently, the default in the code are: 20s to open the session, 15s
to close it.

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=551235
Change-Id: I616a609d4efc320a9f461f810eed37955f20e9a4
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>