[553492] Do not perform Snap Command with ELK Layout
* Since the Snap to Command moves some elements to snap the grid or
other
figures edges, the layout performed by a custom algorithm might be
broken. It is now possible to indicate whether the Snap to is allowed
after the arrange.
* The CustomLayoutAlgorithm has now the information if it allows to
perform a Snap to command after the arrange.
* We retrieve the CustomLayoutAlgorithm matching the
CustomLayoutConfiguration (if present) and we check whether the
algorithm allows to execute a Snap to command after the arrange.

Bug: 553492
Change-Id: I703fe36eb91a38fcc88e6c8c8a03ef4adabcd5bb
Signed-off-by: Florian Barbin <florian.barbin@obeo.fr>