[561385] Revert feature of bug 550382: "Allow more locations for label on border"
This commit reverts the 4 commits concerning this issue ([1], [2], [3]
and [4]).
It also adapts the corresponding migration participant to "revert"
the effects of the new feature and the effects of the corresponding
migration participant.

The data of EditModeTest has been adapted (the double click was not OK
for the test EditModeTest.testShowHideDoubleClickOnNodeLabel). The click
was done on the edge instead of the label. It's strange, it also failed
when the feature was not here.

[1] Revert "[550382] Allow more locations for label on border"
https://git.eclipse.org/c/sirius/org.eclipse.sirius.git/commit/?4185e3b16f15e075f71aa9fe1ea339bf6600913d.
[2] Revert "[550382] Add a migration participant concerning this issue"
https://git.eclipse.org/c/sirius/org.eclipse.sirius.git/commit/?57b9deda512d2610bac72be2470fa4236f09fbf3.
[3] Revert "[550382] Fix migration participant concerning this issue"
https://git.eclipse.org/c/sirius/org.eclipse.sirius.git/commit/?id=5bfd6ddd8c08ed6f39026a8bed8776dc1e4e87ef
[4] Revert "[550382] Add test about allowing more locations for label on
border"
https://git.eclipse.org/c/sirius/org.eclipse.sirius.git/commit/?id=73f3a66dcecbe2bf47d8c5fc6ab5254c895c6548

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=561385
Change-Id: I7791a07d482329e946084bad0aa1384ea5dff9e7
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>