[test] Fix CompartmentsTest when "Comic Sans MS" is not available
The tests of CompartmentsTest depends on the font size. It is not the
same on Windows and Linux, and it depends if "Comic Sans MS" is
installed or not.
This commit adds a tolerance of one pixel to the expected width if the
font is not installed. This is the delta observed on IC server. This
tolerance is added only when it is necessary (on IC server).

Change-Id: If3e0410c822b31d862dfd08e16342e502942e92a
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>