[563730] Modify extension point to declare EClasses as DocumentRoots
Reuse the package_meta_data extension point to add an element to
specify EClasses in the metamodel which should be considered as a
Document Roots.

This is useful for metamodel generated from a XSD. To help Sirius
consider some containment references as not transient even though the
metamodel declares them as transient.

Change-Id: If744a9ab7bb214b665de452982711c947087bfef
Signed-off-by: Stéphane Thibaudeau <stephane.thibaudeau@obeo.fr>
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>