[test] Fix EdgeOnFigureWithAlphaAnchorTest
The test testDragTargetProvidedShapeWithZoomAndScroll fails on some CI
servers (but not on all and not locally on Windows).

This commit changes an assert that was wrong. The test must fail sooner
on CI server. According to the failure screenshot, the dragAndDrop seems
to fail. We can see that several elements are selected instead of only
one, like if the "select + drag" has been considered as a "drag
selection".
This commit also adds:
* a waitUntil(new CheckSelectedCondition()) to be sure that the element
is selected before doing the drag.
* another drag and drop with new start coordinates (in case of failure
of the first). The rounding with the zoom factor is sometimes not what
it should according to the configuration (resolution, font, ...).

Change-Id: I8a53dce0e3bfac47b1b1ee897512f1a31b010e10
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>