[558597] Revert previous changes for Sirius 6.3.1
This reverts commits d3f481ccc, da9865595 and cf72e32b3, which do not
completely fix bug #558597 and cause a regression in the width of
containers with compartments.

We need to release 6.3.1 final in the next few hours for SimRel
2020-06, so the complete fix will be included in a 6.3.2 maintenance
release which should happen in the next few days, but we want to avoid
a partial solution with new regressions to be part of 6.3.1.

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=558597
Change-Id: Iaf78b8ed7b63917624f24784a82d654e137d9829
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>