[564239] Draw combined fragment labels on a OVERLAY layer so that they are always readable
This commit makes the approach of bug 564239 [1] more generic. It is now
also used by the CombinedFragmentEditPart (through the
GradientRoundedRectangle with the field useOverlayLabel set to true) and
not only OperandEditPart (through the OperandFigure).

Even if DDiagramRootEditPart#OVERLAY_LAYER, OverlayLabelsDrawerFigure
and OverlayLabel are used only in sequence diagrams, these classes are
available in standard diagram code for potential further usages.

You can have a look to commit bbac9a8b [2] for details of the initial
implementation.

[1] https://bugs.eclipse.org/bugs/show_bug.cgi?id=560543
[2] https://git.eclipse.org/c/sirius/org.eclipse.sirius.git/commit/?id=bbac9a8b4dbffb277db91cd3991fbc0683e2c141

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=564239
Change-Id: I09b2f7280d821d841b9eac67271121519147ec39
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>