[cleanup] Add the "default" service in the VSM family sample
If this "starter" result is used as starting point for "advanced
tutorial", the service is necessary. So it is better to have it directly
in the "starter result".
Furthermore, this kind of service is automatically created with the
Viewpoint Specification Project wizard so there is no reason to not have
it.

Change-Id: I1aa2de49777f447767734753dea236860576e8c6
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>