[test] Fix ActionDisabledOnSequenceDiagramTest
The test org.eclipse.sirius.tests.swtbot.sequence.AbstractActionDisabledOnSequenceDiagramTest.testPinUnPinWizardFromTabbarOnSequenceDiagramsComponents()
fails on some IC servers when it is executed from
ActionDisabledOnSequenceDiagramTest. In the failure screenshot, we can
see that the editor has not the focus (the Model Explorer view has it).
A waitAllUiEvents is added to be sure that the editor has the focus.

Change-Id: I0cfa820cfb29a112f4d58bc27e24d8d1a612bca2
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>