[565405] Copy deleted message from Eclipse for our own FilteredTree classes
Several changes have been done in Eclipse FilteredTree class. With the
canary target platform, there are some compilation failures, caused by
one of the commit for "bug 564975 - Unused keys in message classes" [1].
The changes in FilteredTree are here since long time but not detected
before.

To be compatible from Eclipse Photon to Eclipse 202009 (canary) we can
not simply duplicate classes from current Eclipse 202009 (as it was done
before for Eclipse Juno to benefit in QuickOutlineControl, of new
features not available in Juno (bugzilla 460712 [2] and commit 302d195d
[3]). Indeed, the delete button of the search field depends on specific
OS implementation, especially Text class for Windows that has been
adapted for that recently ("Bug 293230 [4] - Support SWT.ICON_SEARCH and
SWT.ICON_CANCEL for native find/filter/search box for Windows 7 /
Windows 10").

So this commit simply copies the deleted message and uses it in our own
copies of FilteredTree.

We can switch to the new "implementation" as soon as the bottom
supported version will be 201909.

[1] https://bugs.eclipse.org/bugs/show_bug.cgi?id=564975
[2] https://bugs.eclipse.org/bugs/show_bug.cgi?id=460712
[3] https://git.eclipse.org/c/sirius/org.eclipse.sirius.git/commit/?id=302d195dac44967e5875fc174d43a331ad1f16fe
[4] https://bugs.eclipse.org/bugs/show_bug.cgi?id=293230

Bug: 565405
Change-Id: I3a865c245bce918bc1471896d8a5e654675fe55e
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>