[cleanup] Adapt code to changes made on ELK
The class ElkDiagramLayoutConnector has a compilation error with canary
target platform as the EdgeLabelPlacement.UNDEFINED as been deleted in
ELK. The class ElkDiagramLayoutConnector is initially a copy of
org.eclipse.elk.conn.gmf.GmfDiagramLayoutConnector. This commit does the
same changes as in GmfDiagramLayoutConnector (ie replaces the
EdgeLabelPlacement.UNDEFINED by an Optional#empty()).

Change-Id: I04659263dd915995545d9038d5215e2bdf6504d8
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>