[test] Fix SequenceArrangeLinkedBorderedNodesTest on 202006
Since the activation of 202006 TP on a CI server, the test
org.eclipse.sirius.tests.swtbot.sequence.SequenceArrangeLinkedBorderedNodesTest.test_ArrangeLinkedBorderedNodes()
always fails. The error is "The widget with mnemonic 'All' was not
enabled.". But the real pb is that the editor has not the focus.

This commit:
* adds a call to show method (before setFocus). Sometimes setFocus is
not enough (I've not really understand why).
* adds a "waitUntil(new EditorHasFocusCondition())" to be sure to fail
with the real cause if this test fails again later.

Change-Id: I94b71506e5ef3e8da398da2e46b0fbefe5850b2a
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>