[566778] Avoid reloading the resources during the session closing
At Sirius session closure, the resources must be removed from the
resourceSet once the call of
ResourceStrategyRegistry.getInstance().unloadAtResourceSetDispose on all
the resources of the ResourceSet has been done because otherwise a
resource may be recreated at proxy resolution when calling
unloadAtResourceSetDispose on another one.
For example, if the odesign resource is removed, it is recreated when
calling DRepresentation.getDescription()

A test is needed and will be done in a further commit.

Bug: 566778
Change-Id: If1a8ce25804d802306791978d809c9e7165ec96b
Signed-off-by: Laurent Fasani <laurent.fasani@obeo.fr>