[565885] Change dragTracker associated to some EditPart
The right click behavior has been changed for some EditPart. Before,
when right clicking on the EditPart, it was automatically unselected and
the diagram was selected instead. Now, it is automatically selected and
then the contextual menu is really contextual to the EditPart.

Here are the impacts of the changes. First levels are the classes where
getDragTracker method has been changed and sub levels are the classes
indirectly impacted because they do not implement getDragTracker.
-AbstractSequenceBorderedEditPart
   ExecutionEditPart
     StateEditPArt
-CombinedFragmentCompartmentEditPart and CombinedFragmentEditPart
-OperandCompartmentEditPart and OperandEditPart
-AbstractDiagramBorderNodeEditPart
   DNode2EditPart (border nodes on a node)
     EndOfLifeEditPart
   DNode4EditPart (border nodes on a container)
	 RotativeDNode4EditPart

A test has also been added.

Bug: 565885
Change-Id: If5f23a77d0aa0b3efe96b060d8fa1d88f4bacc3c
Signed-off-by: Laurent Fasani <laurent.fasani@obeo.fr>