[566082] Change the way that default GMF actions are cleaned from Sirius
The previous behavior was not completely OK for Sirius. Indeed, the
default GMF actions are not visible in the menu but they are sometimes
always at the listening of the selection changes of the editor.
This commit removes the declarative part (in plugin.xml) for
menubar="true" and toolbar, and it replaces if by a specific code in
SiriusDiagramActionBarContributor that do the same, removeExistingItem,
but with a dispose of the action in addition.

Bug: 566082
Change-Id: I49a552516f63199f265ad871bb6fe14ff1b66ab4
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>