[566232] Avoid NPE during update of region container layout.
During update of region container layout (can appear after adding
compartment in region container for example), we compare the size of
region
container with the size of its Sirius element. This comparison is done
between an Integer and an int. To avoid NPE when the Integer is null, we
now check if Integer is null before comparison between Integer and int.

Bug: 566232
Change-Id: I5090c8a0c17ba31f8c6cafd1acaffb6fb0d80ede