[567861] Use SiriusPolylineConnectionEx instead of original GMF class
This commit uses SiriusPolylineConnectionEx instead of original GMF
PolylineConnectionEx class.
The class SiriusPolylineConnectionEx has been adapted to have the
correct behavior (styleBits field and associated methods have been
duplicated because of private or final aspects that forbid to correctly
do the fix later).

The method getBounds has been modified to always call the super method.
This is to avoid to "expand" the bounds twice.

Bug: 567861
Change-Id: Ib29451c0bcd623d6078413b6b5013825596a4ac2
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>