[539778] Reverse "layouts order" for ELK layout
Several "arrange all" are called (one for the diagram, and one per
container) at diagram creation (or at diagram opening with new
elements). It seems OK with "classical layout". But it is problematic
for ELK layouts. For ELK, we need to reverse the order of these multiple
"arrange all" from the lowest container level to the highest container
level. For that a new notion of "reverseLayoutsOrder" has been added
into the CustomLayoutAlgorithm.

An issue exists for the "real problem" about why several arrange are
necessary (bugzilla 558016 [1]).

[1] https://bugs.eclipse.org/bugs/show_bug.cgi?id=558016

This commit also avoids to launch an additional and unuseful arrange
command (caused by InitializeHiddenElementsCommand).

Bug: 539778
Change-Id: I6d485d2ae1e0a845b11b42a4b60ff7ca5488c5bd
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>