[568037] Consider compartment as parent of selected elements to arrange
This commit considers AbstractDNode*CompartmentEditPart in
EditPartQuery.getDDiagram().
AbstractDNodeContainerCompartmentEditPart and
AbstractDNodeListCompartmentEditPart are not correctly managed by
EditPartQuery.getDDiagram() as they do not implement
IDiagramElementEditPart. As consequence, the ELK layout was not
retrieved for this element and a standard layout is called instead.

This commit correponds to the comment c3 of the bugzilla:
https://bugs.eclipse.org/bugs/show_bug.cgi?id=568037#c3

This commit also changes bounds of a container and its node from data
needed for EditModeTest. This test uses ELK, I don't know why, but after
this commit, the layout of the container is not done and in the rest of
test, we try to double-click on a node without success. Previously, a
"standard" layout, ie not an ELK one, is done on the contents of the
container.

Bug: 568037
Change-Id: I3be82fc1807873e4b9de451c474ef41a035ba736
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>