[571400] Refresh each InstanceRoleEditPart only once
Each ExecutionEditPart/StateEditPart used to call a refresh on its
parent InstanceRoleEditPart for each single setBounds notification.

On a diagram with 4 lifelines and several hundreds of Execution/State
elements, a Sequence layout or an arrange all would trigger several
hundreds of InstanceRoleEditPart.refresh(). With this patch only one
InstanceRoleEditPart.refresh is triggered per impacted lifeline.

The deleted notificationChanged method was called from the
DiagramEventBroker which is installed before the post commit listener of
the SequenceDiagramEditPart, that's why the
refreshInstanceRoleEditPartsOnAbstractNodeEventSetBounds is called
before the ZOrder refresh.

This commits also reduce the number of decorators refresh.

Bug: 571400
Cherry-picked-from: 567517
Change-Id: I3f6497e2c8eed64f9604c929f411faf95cb77e35
Signed-off-by: Maxime Porhel <maxime.porhel@obeo.fr>