[571400] Create predecessors sets
The first commit done for this issue (commit
0f974759277c2b1a65154c68fb4fad25124f5978) replaced calls to
List.retainAll calls during layout with Sets.intersections with measured
performance improvements for SequenceLayout used in arrange all, diagram
opening and precommit layout.

But in some cases, when there is no graphical predecessors, we need to
inspect the flagged predecessors, and the implementation has performance
issues: the Sets.intersection returns a SetView and the effective
intersection will be computed for each calls to size() or each iteration
using a new iterator().

Furthermore, we need to have access to the size of the semantic and
graphical predecessors and then to iterate on both, and if there is a
desynchronization, we need to get the last flagged predecessor which has
the same semantic predecessor index.

We need to force the computation of the semantic and graphical
predecessors sets but as the while loop does not need to access the
size() of the flagged ends predecessors, we can keep this sets with the
lazy computation brought by Guava.

Bug: 571400
Cherry-picked-from: 567517
Change-Id: I37f3debad39788c0da9d2d373c4b6375f1e299cd
Signed-off-by: Maxime Porhel <maxime.porhel@obeo.fr>