[553497] ELK - Improvement concerning "arrange at opening"
This commit adds an information to differentiate arrange at opening and
an arrange selection launched by the end-user.

This commit also changes bounds of a container from data
needed for EditModeTest. This test uses ELK, I don't know why, but after
this commit, the layout of the container at opening is different and
causes overlaps and in the rest of test, we try to double-click on a
node without success.

Bug: 553497
Change-Id: Idd54162172141b4e25a5a78144edf6e3632e91b0
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>