[test] Make DeleteFromDiagramTest more reliable
Sometimes, on some CI servers, the test
DeleteFromDiagramTest.testDeleteFromDiagramAction() fails. The error was
java.lang.AssertionError: expected:<50> but was:<0>
	at org.eclipse.sirius.tests.swtbot.DeleteFromDiagramTest.testDeleteFromDiagramAction(DeleteFromDiagramTest.java:83)

It was like if the SWTBotTree in
AbstractSiriusSwtBotGefTestCase.getNbStatusInErrorLog() has not the time
to be fill before the method "tree.getAllItems().length" was called.

To avoid this kind of problem, the way to detect new messages has been
changed. It now relies on the existing "catch errors" and "catch
messages" mechanism (instead of relying on UI).

Change-Id: Ida0b9e6ac874e300f3e4c5f9c2e5f82908c14d33
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>