[cleanup] Ensure a predictable iteration order for Session views getters
Use a LinkedHashSet instead of an HashSet for Session.getOwnedViews()
and Session.getSeelctedViews(boolean) as HashSet does not provide "
guarantees as to the iteration order of the set; in particular, it does
not guarantee that the order will remain constant over time."

The issue has been seen on production with a random issue specific to a
client product (linked to the ViewpointRegistry instanciation in this
product). The bug was random due to this HashSet implementation which
was hiding the concrete issue.

Change-Id: Idcfa7c8b76fa61e4158f506058de6c32530ebfdf
Signed-off-by: Maxime Porhel <maxime.porhel@obeo.fr>