[568572] Fix invalid representation status computation
If the diagram URI prefix indicates that the representation is stored in
an aird, it does not necessarily imply that "if the representation
exists then it would be yet loaded then we can rely on
isLoadedRepresentation". Indeed, if the representation has been moved
from a resource to another it may not be loaded yet and it is though
reachable.

Bug: 568572
Change-Id: I8acfd22698b99b556a74393a3912df95b5fd5608
Signed-off-by: Laurent Fasani <laurent.fasani@obeo.fr>