[570366] Add adapter in DRepresentations to reference its descriptor
The process creating a representation and its descriptor has been
updated in order to link them as soon as possible. This way the
descriptor is available when a DAnalisysSelector is processing the
representation location.
SiriusControlAndCrossReferenceTest.testControl() was updated in order to
check that when a DAnalysisSelector is processing a new representation,
its name is accessible. This was a regression due to commit
ef6f0f80f3f5211e1e52a1e6b2bdb46c1ec356d2.

Bug: 570366
Change-Id: I6d58465f22d3425104521a739fb684a7acbf5dd5
Signed-off-by: Steve Monnier <steve.monnier@obeo.fr>