[570055] Avoid to adapt node label when it is on border
There is no reason to change the location and size of a label if this
last is on the border of its parent node. This change seems not
mandatory. Indeed, a recent change, commit 753dc1ed [1], has a side
effect to also fix this problem. But the analysis of this issue shows
that the code in
org.eclipse.sirius.diagram.ui.tools.api.graphical.edit.styles.SimpleStyleConfiguration.adaptNodeLabel(DNode,
SiriusWrapLabel) is not necessary for label on border.

The test IconRefreshTests has been updated to reveal the problem. The
test fails without this fix and without the commit [1].
For that, in this commit:
* The VSM has been updated to add a conditional style to reveal the
problem.
* The VSM has been updated to add a Node with label on border to confirm
that the problem is only for border node with label on border.
* The semantic model has been fixed (it was wrong).
* The diagram of the aird file has just been refreshed.

[1] https://git.eclipse.org/c/sirius/org.eclipse.sirius.git/commit/?h=v6.4.x&id=753dc1ed3a8a1d768c63f3d4618356efc017ccbd

Bug: 570055
Change-Id: I6ee25c3c455d81e11b0b8a2ae979543f8490bc12
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>