[571360] Change operations used to modify models during copy format
Some operations do not triggered a notification or to be more precise
the notification triggered is not handled by the EditPart (as is is not
expected by GMF). This is 2 cases:
* NotationPackage.Literals.NODE__LAYOUT_CONSTRAINT, only SIZE__WIDTH,
SIZE__HEIGH, LOCATION__X and LOCATION__Y are considered: for this case,
we adapt the SiriusTextEditPart and SiriusNoteEditPart to consider the
NODE__LAYOUT_CONSTRAINT.
* NotationPackage.Literals.EDGE__BENDPOINTS, only
NotationPackage.Literals.RELATIVE_BENDPOINTS__POINTS is considered: for
this case, we change the way to set the bendpoints.

Bug: 571360
Cherry-picked-from: 571299
Change-Id: I4a1769d466a4dce2f29ad51ad4994021d4c94c0e
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>