[571360] Do not reset label constraint on font change
There is no reason to reset the label constraint to figure constraint
during a call to method refreshFont. Furthermore, this can, in some
cases, override the previous constraint set according to a location
change.
Moreover, the deleted code was wrong. The location set as constraint
into the DBorderItemLocator was in absolute coordinates instead of
relative coordinates.

Bug: 571360
Cherry-picked-from: 571299
Change-Id: Ida35556c22f63bb96148ea54bd09fce5789155ff
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>