[571304] Copy the format of label on border in sequence diagram
The method
org.eclipse.sirius.diagram.ui.business.api.view.SiriusGMFHelper.getLabelNode(View)
has been update. Indeed, with the copy/paste API, the label is not
always in first position of the children list.

The data has been regenerated. Even if there is no label displayed in
sequence tests, the VSM describes a label, not displayed, on border of
Execution and on Lifeline. So the model is impacted even if the
displayed diagram is not.

Bug: 571304
Change-Id: I4cff0a4f24e7e1703b206f7e71952d8a9fab35f3
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>