[571115] Blank regeneration of the VSM table editor
In this commit, the copyright has been changed in table.genmodel,
because when Pierre-Charles modified it in commit efc8b2ad [1], he has
partially changed the line ending format. In consequence, the format is
wrong when the generation is done on Windows.
Furthermore, the regen of editor has not been done after the
modification of the Copyright by Pierre in commit 75ddffc0 [2].

* Regen done with Acceleo 2.8.0
* Regen done without modification in MM
* 2 clean-up have been launched after regen

[1] https://git.eclipse.org/c/sirius/org.eclipse.sirius.git/commit/?id=efc8b2ad33e3463ec313769e7d2823e928db35da
[2] https://git.eclipse.org/c/sirius/org.eclipse.sirius.git/commit/?id=75ddffc002cd3db43dd1808d5f4385616a242cfe

Bug: 571115
Change-Id: I70e0129e57681a9d704f33a50d6d4b25e6a093b3
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>