[571115] Add edition tests using specific CellEditors
The CellEditionTest data has been refactored to allow several kind of
edition (and not only for VP-2683 old case).

This commit needs an improvement in SwtBot (bugzilla 571838) [1].
Currently this improvement is not integrated in SwtBot and Sirius uses
an "old" 2.8 version of GMF. The classes SWTBotTree and SWTBotTreeItem
have been copied in Sirius to benefit of the improvement. The original
classes have been copied from commit corresponding to 2.8.0 version [2].

[1] https://bugs.eclipse.org/bugs/show_bug.cgi?id=571838
[2] https://git.eclipse.org/c/swtbot/org.eclipse.swtbot.git/commit/?id=2377612904ceab6a49a81b127a6573da5b88a5c8


Bug: 571115
Change-Id: Ic7788d21c2fc1d8f3eb8e3be09a8c294801ae672
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>