[571287] Avoid potential SWTException exception: "Widget is disposed"
This commit fixes the problem seen during validation (comment 7 [1] of
the issue):
* To avoid the "Widget is disposed" problem that occurs during the
UpdateRequest runnable launched in async by the DeferredUpdateManager,
the executions of them are forced before cleaning a the offscreen
diagram EditPart.
* The dispose of the shell is now done in
MappingBasedSiriusFormatManagerFactory.cleanAndDispose(DiagramEditPart).
Indeed, running it in async caused a problem with the above change.
* All the diagramEditPart are disposed. This is a not a problem detected
but a potential problem seen during analysis.

[1] https://bugs.eclipse.org/bugs/show_bug.cgi?id=571249#c7

Bug: 571287
Cherry-picked-from: 571249
Change-Id: Ib58067fd0cdceb77d42fd9f53ab4118a5525e4dd
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>