[571287] Regenerate data of tests
* The edge layout of "DiagType2 with notes of MyPackage" has been
changed as the GMF bendpoints are wrong in aird file. This is not really
a problem but it was confusing when we try to understand the problem.

The regen was done previously without MB_GENERATE_IMAGES_TEST_DATA and
the result is different for edge layout (especially the list of draw2d
points). It seems that the recorded data are wrong concerning edge
layout data. The draw2d points list correspond to a wrong layout. I'm
not sure about what is the problem: a problem to look after, but later.
One thing is that with the fix of bugzill 571249, the result with
MB_GENERATE_IMAGES_TEST_DATA to true is expected. Indeed, when generated
with MB_GENERATE_IMAGES_TEST_DATA, a
EclipseUIUtil.synchronizeWithUIThread is done during
DiagramDialectUIServices.exportWithResult.

Bug: 571287
Cherry-picked-from: 571249
Change-Id: Ia9ddc55d19d0d2e5abb289bca0e006f24bb8be99
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>